figure out how to persist station data manually once
